Refactoring of the original server/node code. Renamed client to node to be a more accurate description. See after the break for more details.
Current Functionality: node.py can be called via the GET url /node/register which will contact the Master node and register its IP and PORT in the new Grid object, Master then returns the result of this attempted action to the Node. The Node then parses this response and outputs its own response to the user. This actual functionality is unlikely to actually use the webserver for rendering in real life, this is just a demo.

- Abstracted a new http layer to handle the requests and responses including their conversion to JSON
- Renamed client.py to node.py as this is a more accurate description, we will later have a client.py and that would just get really confusing.
- Added a new Grid class that will eventually be refactored around, kept in memory and will store the state of the grid.
- Added a working node_register_POST function to server that actually adds a sent valid JSON request of a node to the Grid.
- Other stuff, read the diff if you care that much :P