Server/Node back and forth completed. Node creation of processes completed. See extended description for more details.
This was a long commit that would ideally be smaller ones however as I
was changing a number of small things at the same time they all kind of
got lost in the overall getting things to work process, so here is a
list of things done.

- Nodes will now automatically detect how many cores are available
- Nodes can now receive a Task which is a Node representation of a Work
Unit
- Nodes can now take a Task and execute it, monitor for its completion,
then inform the Server.
- Nodes, while executing a task, write the stdout and stderr to files
and reads in from stdin
- Nodes now have two threads. One for the heartbeat, and one for the
process monitor.
- Modified the work unit dict transform to include more information
about the job so that requesting a specific work unit isn't totally
devoid of information relevant to that work unit such as the
executable, flags and walltime
- Renamed command to flags, command is now a dynamic property that
combines the executable and the flags

To Do:
- Send input files to Node along with the Task
- Send output files to the Server when complete
- Store output files on Node in a proper location
- Clean up output files on Node once they have been sent to the Server
- Remove nodes from the list when they go offline
- Make it possible to dynamically change the Scheduler used
- Properly handle the case of failing to connect to the Node when
attempting to send the Task
- Restructure jobs such that they can handle the fact that files wont
be available immediately (Haven't decided how best to do this yet)
- Authentication Layer
- Tidy up and reorganise the code as it's gotten a little messy at the
moment
- Rename views to controllers as they should be