Moved job output exception handling into The Grid controller. Made killing jobs more sophisticated.
When requesting job output an info_msg is passed back regarding the status of the job.
the list of URIs contains job files for finished and killed work units (if they exist).

The Grid now differentiates between killed jobs and killed work units.
Jobs are marked as finished if all work units are either finished or killed.
Jobs are marked as killed if a kill request comes from client.
 - When this happens a kill_msg is stored in the job
Work units are marked as killed if the Node kills the job.
 - When this happens a kill_msg is stored in the work unit

When requesting the output of a job, warnings are given for the job/individual
work units if they are killed saying that the output will be incomplete.
Kill message for the job and individual work units is reported back to the client as part of the info_msg.